+++ Pong/Player.h
@@ -2,10 +2,10 @@
 #include <SFML/Graphics.hpp>
 
 const float PLAYER_SPEED = 0.0006f;
-const int32_t PLAYER_WIDTH = 18;
-const int32_t PLAYER_HEIGHT = 104;
-const int32_t HALF_PLAYER_WIDTH = PLAYER_WIDTH / 2;
-const int32_t HALF_PLAYER_HEIGHT = PLAYER_HEIGHT / 2;
+const uint16_t PLAYER_WIDTH = 18;
+const uint16_t PLAYER_HEIGHT = 104;
+const uint16_t HALF_PLAYER_WIDTH = PLAYER_WIDTH / 2;
+const uint16_t HALF_PLAYER_HEIGHT = PLAYER_HEIGHT / 2;
 
 class Game;
 class Ball;
@@ -15,12 +15,11 @@ public:
     friend Game;
     friend Ball;
 
-    Player(float x, float y, int32_t id);
+    Player(float x, float y, int8_t id);
     void Update(float time);
-
 private:
     float x, y;
-    int32_t id, score;
+    int8_t id, score;
     sf::Sprite sprite;
     sf::Texture texture;
